+++ {{cookiecutter.repo_name}}/project/mods_available/contact_form/settings.py
@@ -1,9 +1,9 @@
 # -*- coding: utf-8 -*-
 
 INSTALLED_APPS = add_to_tuple(INSTALLED_APPS,
-    'captcha',
     'import_export',
     'emencia.django.countries',
-    'project.contact_form')
+    'project.contact_form',
+)
 
 CONTACT_FORM_TO = []