+++ {{cookiecutter.repo_name}}/project/mods_available/socialaggregator/__init__.py
@@ -5,4 +5,7 @@ Enable the `emencia-django-socialaggregator`_ app to manage social contents.
 
 .. note::
         This app require to fill some API keys settings (like for Twitter API, Facebook API, etc..) to work correctly.
+
+.. warning::
+        This mod is deprecated since socialaggregator has not been ported to ``Django>=1.7``
 """